test: consolidate unit test suite
