+++ tests/SOURCE_GUARDS.md
@@ -8,7 +8,7 @@ extraction.
 | Guard file | Invariant label | Owning domain | Why behavior coverage is insufficient | Removal condition |
 | --- | --- | --- | --- | --- |
 | `tests/unit/admin/slides-actions.test.ts` | `admin-slides-enhance-order` | Admin/slides UI | The invariant is Svelte form-enhance callback ordering in `+page.svelte`; Bun has no DOM runner in this suite to exercise the editor failure path. | Replace with a browser/component test that submits the create/update custom slide form and proves failure preserves editor content while success closes after `update()`. |
-| `tests/unit/auth/routes.test.ts` | `auth-returnTo-threading` | Auth routing | Pure validator behavior is covered, but the redirect page/server data threading spans SvelteKit route files that are not all callable as isolated Bun units. | Add route/component harness coverage for the redirect load/page flow where onboarding context wins and safe `returnTo` is used otherwise. |
+| `tests/unit/auth/core.test.ts` | `auth-returnTo-threading` | Auth routing | Pure validator behavior is covered, but the redirect page/server data threading spans SvelteKit route files that are not all callable as isolated Bun units. | Add route/component harness coverage for the redirect load/page flow where onboarding context wins and safe `returnTo` is used otherwise. |
 | `tests/unit/security/basics.test.ts` | `hooks-anon-admin-returnTo` | Auth hooks | `authorizationHandle` is composed inside `hooks.server.ts` and not exported separately; source guard pins anonymous admin path encoding and non-admin dashboard redirect. | Export a behavior-neutral authorization helper or add hook-level request tests that can invoke this branch directly. |
 | `tests/unit/security/basics.test.ts` | `onboarding-bootstrap-origin` | Onboarding hooks | The bootstrap banner call is an initialization side effect inside the hook chain; the negative origin argument is not externally observable in Bun unit tests. | Extract the banner invocation decision or add a hook harness that spies on the banner call without importing request-derived origin. |
 | `tests/unit/onboarding/settings-substep-walk.test.ts` | `onboarding-settings-substeps` | Onboarding UI | The multi-substep footer/rendering behavior lives in a Svelte page and requires DOM interaction for true behavior coverage. | Replace with component/browser tests that walk the settings substeps and assert submit button gating. |
