+++ tests/unit/admin/csrf-action.test.ts
@@ -6,14 +6,13 @@ import {
 	getAppSetting,
 	setAppSetting
 } from '$lib/server/admin/settings.service';
-import { db } from '$lib/server/db/client';
-import { appSettings } from '$lib/server/db/schema';
 // Re-pointed to the nested Security route after US-022 deleted the monolith.
 // All 14 tests below exercise the same updateCsrfOrigin contract — the
 // nested-route handler is a verbatim copy of the monolith's (per the
 // commit 220f6ad CsrfOrigin inline OCC retrofit, then carried forward
 // when the Security tab was extracted in 97152be).
 import { actions } from '../../../src/routes/admin/settings/security/+page.server';
+import { resetSharedTestDb } from '../../helpers/db';
 
 const ORIGIN = 'http://localhost:5173';
 
@@ -63,9 +62,7 @@ async function runUpdate(request: Request, urlOverride?: string) {
 }
 
 describe('admin updateCsrfOrigin action', () => {
-	beforeEach(async () => {
-		await db.delete(appSettings);
-	});
+	beforeEach(resetSharedTestDb);
 
 	it('returns fail(409) and does NOT call setAppSetting when origin mismatches and confirmMismatch is absent', async () => {
 		const result = (await runUpdate(
