+++ tests/unit/stats/serialization.test.ts
@@ -15,410 +15,149 @@ import {
 	type UserStats
 } from '$lib/server/stats/types';
 
-/**
- * Unit tests for Stats Serialization
- *
- * Tests JSON serialization, parsing, type detection, and round-trip validation.
- */
+const ranked = [{ rank: 1, title: 'Example', count: 5, thumb: null }];
+const distribution = (length: number, value: number) => ({
+	minutes: Array(length).fill(value),
+	plays: Array(length).fill(1)
+});
+const commonStats = {
+	year: 2024,
+	totalWatchTimeMinutes: 6000,
+	totalPlays: 200,
+	topMovies: ranked,
+	topShows: ranked,
+	topGenres: ranked,
+	watchTimeByMonth: distribution(12, 500),
+	watchTimeByHour: distribution(24, 250),
+	watchTimeByWeekday: distribution(7, 857),
+	contentTypes: {
+		movies: { count: 50, minutes: 3000 },
+		episodes: { count: 150, minutes: 3000 },
+		tracks: { count: 0, minutes: 0 }
+	},
+	decadeDistribution: [],
+	seriesCompletion: [],
+	topRewatches: [],
+	marathonDay: null,
+	watchStreak: null,
+	yearComparison: null,
+	longestBinge: null,
+	firstWatch: null,
+	lastWatch: null
+} satisfies Omit<UserStats, 'userId' | 'percentileRank'>;
 
-// =============================================================================
-// Test Helpers
-// =============================================================================
+function userStats(overrides: Partial<UserStats> = {}): UserStats {
+	return { ...commonStats, userId: 1, percentileRank: 85, ...overrides };
+}
 
-function createValidUserStats(overrides: Partial<UserStats> = {}): UserStats {
+function serverStats(overrides: Partial<ServerStats> = {}): ServerStats {
 	return {
-		userId: 1,
-		year: 2024,
-		totalWatchTimeMinutes: 6000,
-		totalPlays: 200,
-		topMovies: [{ rank: 1, title: 'The Matrix', count: 5, thumb: null }],
-		topShows: [{ rank: 1, title: 'Breaking Bad', count: 50, thumb: null }],
-		topGenres: [{ rank: 1, title: 'Action', count: 100, thumb: null }],
-		watchTimeByMonth: {
-			minutes: [500, 400, 600, 500, 400, 300, 800, 600, 500, 400, 500, 500],
-			plays: [10, 8, 12, 10, 8, 6, 16, 12, 10, 8, 10, 10]
-		},
-		watchTimeByHour: { minutes: Array(24).fill(250), plays: Array(24).fill(5) },
-		watchTimeByWeekday: { minutes: Array(7).fill(857), plays: Array(7).fill(29) },
-		contentTypes: {
-			movies: { count: 50, minutes: 3000 },
-			episodes: { count: 150, minutes: 3000 },
-			tracks: { count: 0, minutes: 0 }
-		},
-		decadeDistribution: [],
-		seriesCompletion: [],
-		topRewatches: [],
-		marathonDay: null,
-		watchStreak: null,
-		yearComparison: null,
-		percentileRank: 85,
-		longestBinge: null,
-		firstWatch: null,
-		lastWatch: null,
+		...commonStats,
+		totalUsers: 10,
+		topViewers: [{ rank: 1, userId: 1, username: 'TopUser', totalMinutes: 10000 }],
 		...overrides
 	};
 }
 
-function createValidServerStats(overrides: Partial<ServerStats> = {}): ServerStats {
-	return {
-		year: 2024,
-		totalUsers: 10,
-		totalWatchTimeMinutes: 60000,
-		totalPlays: 2000,
-		topMovies: [{ rank: 1, title: 'Popular Movie', count: 50, thumb: null }],
-		topShows: [{ rank: 1, title: 'Popular Show', count: 500, thumb: null }],
-		topGenres: [{ rank: 1, title: 'Drama', count: 500, thumb: null }],
-		watchTimeByMonth: { minutes: Array(12).fill(5000), plays: Array(12).fill(100) },
-		watchTimeByHour: { minutes: Array(24).fill(2500), plays: Array(24).fill(50) },
-		watchTimeByWeekday: { minutes: Array(7).fill(8571), plays: Array(7).fill(286) },
-		contentTypes: {
-			movies: { count: 500, minutes: 30000 },
-			episodes: { count: 1500, minutes: 30000 },
-			tracks: { count: 0, minutes: 0 }
-		},
-		decadeDistribution: [],
-		seriesCompletion: [],
-		topRewatches: [],
-		marathonDay: null,
-		watchStreak: null,
-		yearComparison: null,
-		topViewers: [{ rank: 1, userId: 1, username: 'TopUser', totalMinutes: 10000 }],
-		longestBinge: null,
-		firstWatch: null,
-		lastWatch: null,
-		...overrides
-	};
+function expectStatsParseError(action: () => unknown, message?: string) {
+	try {
+		action();
+		expect.unreachable('Expected StatsParseError');
+	} catch (error) {
+		expect(error).toBeInstanceOf(StatsParseError);
+		if (message) expect((error as StatsParseError).message).toContain(message);
+	}
 }
 
 describe('stats type guards', () => {
 	it.each([
-		['user stats', isUserStats, createValidUserStats(), true],
-		[
-			'user-shaped stats with custom identity',
-			isUserStats,
-			createValidUserStats({ userId: 42 }),
-			true
-		],
-		['server stats as user stats', isUserStats, createValidServerStats(), false],
-		['server stats', isServerStats, createValidServerStats({ totalUsers: 25 }), true],
-		['user stats as server stats', isServerStats, createValidUserStats(), false]
+		['user stats', isUserStats, userStats(), true],
+		['server stats as user stats', isUserStats, serverStats(), false],
+		['server stats', isServerStats, serverStats({ totalUsers: 25 }), true],
+		['user stats as server stats', isServerStats, userStats(), false]
 	] as const)('%s -> %s', (_name, guard, stats, expected) => {
 		expect(guard(stats)).toBe(expected);
 	});
 
 	it.each([
-		[createValidUserStats({ totalPlays: 1, totalWatchTimeMinutes: 0 }), true],
-		[createValidServerStats({ totalPlays: 0, totalWatchTimeMinutes: 30 }), true],
-		[createValidUserStats({ totalPlays: 0, totalWatchTimeMinutes: 0 }), false]
+		[userStats({ totalPlays: 1, totalWatchTimeMinutes: 0 }), true],
+		[serverStats({ totalPlays: 0, totalWatchTimeMinutes: 30 }), true],
+		[userStats({ totalPlays: 0, totalWatchTimeMinutes: 0 }), false]
 	] as const)('hasWatchHistory(%p) -> %s', (stats, expected) => {
 		expect(hasWatchHistory(stats)).toBe(expected);
 	});
 });
 
-describe('Stats Serialization', () => {
-	// =========================================================================
-	// serializeStats
-	// =========================================================================
+describe('stats serialization contracts', () => {
+	it.each([
+		['user', userStats(), parseUserStats, 'userId', 1],
+		['server', serverStats(), parseServerStats, 'totalUsers', 10]
+	] as const)('serializes, parses, auto-detects, and round-trips %s stats', (_name, stats, parser, key, value) => {
+		const json = serializeStats(stats);
+		const parsed = parser(json);
+		const detected = parseStats(json);
 
-	describe('serializeStats', () => {
-		it('serializes UserStats to JSON string', () => {
-			const stats = createValidUserStats();
-			const json = serializeStats(stats);
-
-			expect(typeof json).toBe('string');
-			expect(json).toContain('"userId":1');
-			expect(json).toContain('"year":2024');
-		});
-
-		it('serializes ServerStats to JSON string', () => {
-			const stats = createValidServerStats();
-			const json = serializeStats(stats);
-
-			expect(typeof json).toBe('string');
-			expect(json).toContain('"totalUsers":10');
-			expect(json).toContain('"year":2024');
-		});
+		expect(json).toContain(`"${key}":${value}`);
+		expect(parsed).toEqual(stats);
+		expect(detected).toEqual(stats);
+		expect(roundTripStats(stats)).toEqual(stats);
 	});
 
-	// =========================================================================
-	// parseUserStats
-	// =========================================================================
-
-	describe('parseUserStats', () => {
-		it('parses valid UserStats JSON', () => {
-			const stats = createValidUserStats();
-			const json = JSON.stringify(stats);
-
-			const parsed = parseUserStats(json);
-
-			expect(parsed.userId).toBe(1);
-			expect(parsed.year).toBe(2024);
-			expect(parsed.totalWatchTimeMinutes).toBe(6000);
+	it('preserves nullable rich fields through round-trip', () => {
+		const original = userStats({
+			longestBinge: { startTime: 1704067200, endTime: 1704085200, plays: 6, totalMinutes: 300 },
+			firstWatch: { title: 'First Movie', viewedAt: 1704067200, thumb: '/thumb.jpg', type: 'movie' }
 		});
 
-		it('throws StatsParseError for invalid JSON', () => {
-			try {
-				parseUserStats('not valid json');
-				expect.unreachable('Should have thrown');
-			} catch (error) {
-				expect(error).toBeInstanceOf(StatsParseError);
-				expect((error as StatsParseError).message).toContain('Invalid JSON string');
-			}
-		});
-
-		it('throws StatsParseError for invalid UserStats structure', () => {
-			const invalidStats = { notAUserStats: true };
-
-			try {
-				parseUserStats(JSON.stringify(invalidStats));
-				expect.unreachable('Should have thrown');
-			} catch (error) {
-				expect(error).toBeInstanceOf(StatsParseError);
-				expect((error as StatsParseError).message).toContain('Invalid UserStats');
-			}
-		});
-
-		it('throws StatsParseError for missing required field', () => {
-			const stats = createValidUserStats();
-			const { userId, ...statsWithoutUserId } = stats;
-
-			try {
-				parseUserStats(JSON.stringify(statsWithoutUserId));
-				expect.unreachable('Should have thrown');
-			} catch (error) {
-				expect(error).toBeInstanceOf(StatsParseError);
-			}
-		});
+		expect(roundTripStats(original)).toEqual(original);
 	});
 
-	// =========================================================================
-	// parseServerStats
-	// =========================================================================
-
-	describe('parseServerStats', () => {
-		it('parses valid ServerStats JSON', () => {
-			const stats = createValidServerStats();
-			const json = JSON.stringify(stats);
-
-			const parsed = parseServerStats(json);
-
-			expect(parsed.totalUsers).toBe(10);
-			expect(parsed.year).toBe(2024);
-		});
-
-		it('throws StatsParseError for invalid JSON', () => {
-			try {
-				parseServerStats('{invalid');
-				expect.unreachable('Should have thrown');
-			} catch (error) {
-				expect(error).toBeInstanceOf(StatsParseError);
-				expect((error as StatsParseError).message).toContain('Invalid JSON string');
-			}
-		});
-
-		it('throws StatsParseError for invalid ServerStats structure', () => {
-			const invalidStats = { someField: 'value' };
-
-			try {
-				parseServerStats(JSON.stringify(invalidStats));
-				expect.unreachable('Should have thrown');
-			} catch (error) {
-				expect(error).toBeInstanceOf(StatsParseError);
-				expect((error as StatsParseError).message).toContain('Invalid ServerStats');
-			}
-		});
+	it.each([
+		['parseUserStats invalid JSON', () => parseUserStats('not valid json'), 'Invalid JSON string'],
+		[
+			'parseUserStats invalid structure',
+			() => parseUserStats(JSON.stringify({ ...userStats(), userId: undefined })),
+			'Invalid UserStats'
+		],
+		['parseServerStats invalid JSON', () => parseServerStats('{invalid'), 'Invalid JSON string'],
+		[
+			'parseServerStats invalid structure',
+			() => parseServerStats(JSON.stringify({ someField: 'value' })),
+			'Invalid ServerStats'
+		],
+		['parseStats invalid JSON', () => parseStats('not json'), 'Invalid JSON string'],
+		[
+			'parseStats ambiguous object',
+			() => parseStats(JSON.stringify({ year: 2024 })),
+			'Unable to determine stats type'
+		],
+		[
+			'parseStats array',
+			() => parseStats(JSON.stringify([1, 2, 3])),
+			'Unable to determine stats type'
+		],
+		['parseStats null', () => parseStats('null'), 'Unable to determine stats type'],
+		['parseStats string', () => parseStats('"just a string"'), 'Unable to determine stats type'],
+		['parseStats number', () => parseStats('42'), 'Unable to determine stats type'],
+		[
+			'parseStats invalid user branch',
+			() => parseStats(JSON.stringify({ ...userStats(), userId: 'not a number' })),
+			'Invalid UserStats'
+		],
+		[
+			'parseStats invalid server branch',
+			() => parseStats(JSON.stringify({ ...serverStats(), totalUsers: 'not a number' })),
+			'Invalid ServerStats'
+		]
+	] as const)('%s throws StatsParseError', (_name, action, message) => {
+		expectStatsParseError(action, message);
 	});
 
-	// =========================================================================
-	// parseStats (auto-detection)
-	// =========================================================================
-
-	describe('parseStats', () => {
-		it('detects and parses UserStats based on userId field', () => {
-			const stats = createValidUserStats();
-			const json = JSON.stringify(stats);
-
-			const parsed = parseStats(json);
-
-			expect('userId' in parsed).toBe(true);
-			expect((parsed as UserStats).userId).toBe(1);
-		});
-
-		it('detects and parses ServerStats based on totalUsers field', () => {
-			const stats = createValidServerStats();
-			const json = JSON.stringify(stats);
-
-			const parsed = parseStats(json);
-
-			expect('totalUsers' in parsed).toBe(true);
-			expect((parsed as ServerStats).totalUsers).toBe(10);
-		});
-
-		it('throws StatsParseError for invalid JSON', () => {
-			try {
-				parseStats('not json');
-				expect.unreachable('Should have thrown');
-			} catch (error) {
-				expect(error).toBeInstanceOf(StatsParseError);
-				expect((error as StatsParseError).message).toContain('Invalid JSON string');
-			}
-		});
-
-		it('throws StatsParseError when neither userId nor totalUsers present', () => {
-			const ambiguousData = {
-				year: 2024,
-				totalWatchTimeMinutes: 1000,
-				totalPlays: 50,
-				topMovies: [],
-				topShows: [],
-				topGenres: [],
-				watchTimeByMonth: { minutes: Array(12).fill(0), plays: Array(12).fill(0) },
-				watchTimeByHour: { minutes: Array(24).fill(0), plays: Array(24).fill(0) },
-				longestBinge: null,
-				firstWatch: null,
-				lastWatch: null
-			};
-
-			try {
-				parseStats(JSON.stringify(ambiguousData));
-				expect.unreachable('Should have thrown');
-			} catch (error) {
-				expect(error).toBeInstanceOf(StatsParseError);
-				expect((error as StatsParseError).message).toContain('Unable to determine stats type');
-			}
-		});
-
-		it('throws StatsParseError for non-object JSON (array)', () => {
-			try {
-				parseStats(JSON.stringify([1, 2, 3]));
-				expect.unreachable('Should have thrown');
-			} catch (error) {
-				expect(error).toBeInstanceOf(StatsParseError);
-				expect((error as StatsParseError).message).toContain('Unable to determine stats type');
-			}
-		});
-
-		it('throws StatsParseError for null JSON', () => {
-			try {
-				parseStats('null');
-				expect.unreachable('Should have thrown');
-			} catch (error) {
-				expect(error).toBeInstanceOf(StatsParseError);
-				expect((error as StatsParseError).message).toContain('Unable to determine stats type');
-			}
-		});
-
-		it('throws StatsParseError for primitive JSON (string)', () => {
-			try {
-				parseStats('"just a string"');
-				expect.unreachable('Should have thrown');
-			} catch (error) {
-				expect(error).toBeInstanceOf(StatsParseError);
-				expect((error as StatsParseError).message).toContain('Unable to determine stats type');
-			}
-		});
-
-		it('throws StatsParseError for primitive JSON (number)', () => {
-			try {
-				parseStats('42');
-				expect.unreachable('Should have thrown');
-			} catch (error) {
-				expect(error).toBeInstanceOf(StatsParseError);
-				expect((error as StatsParseError).message).toContain('Unable to determine stats type');
-			}
-		});
-
-		it('throws StatsParseError when userId present but validation fails', () => {
-			const invalidUserStats = {
-				userId: 'not a number', // Should be number
-				year: 2024
-			};
-
-			try {
-				parseStats(JSON.stringify(invalidUserStats));
-				expect.unreachable('Should have thrown');
-			} catch (error) {
-				expect(error).toBeInstanceOf(StatsParseError);
-				expect((error as StatsParseError).message).toContain('Invalid UserStats');
-			}
-		});
-
-		it('throws StatsParseError when totalUsers present but validation fails', () => {
-			const invalidServerStats = {
-				totalUsers: 'not a number', // Should be number
-				year: 2024
-			};
-
-			try {
-				parseStats(JSON.stringify(invalidServerStats));
-				expect.unreachable('Should have thrown');
-			} catch (error) {
-				expect(error).toBeInstanceOf(StatsParseError);
-				expect((error as StatsParseError).message).toContain('Invalid ServerStats');
-			}
-		});
-	});
-
-	// =========================================================================
-	// roundTripStats
-	// =========================================================================
-
-	describe('roundTripStats', () => {
-		it('round-trips UserStats correctly', () => {
-			const original = createValidUserStats();
-			const roundTripped = roundTripStats(original);
-
-			expect(roundTripped).toEqual(original);
-		});
-
-		it('round-trips ServerStats correctly', () => {
-			const original = createValidServerStats();
-			const roundTripped = roundTripStats(original);
-
-			expect(roundTripped).toEqual(original);
-		});
-
-		it('preserves all fields through round-trip', () => {
-			const original = createValidUserStats();
-			original.longestBinge = {
-				startTime: 1704067200,
-				endTime: 1704085200,
-				plays: 6,
-				totalMinutes: 300
-			};
-			original.firstWatch = {
-				title: 'First Movie',
-				viewedAt: 1704067200,
-				thumb: '/thumb.jpg',
-				type: 'movie'
-			};
-
-			const roundTripped = roundTripStats(original) as UserStats;
-
-			expect(roundTripped.longestBinge).toEqual(original.longestBinge);
-			expect(roundTripped.firstWatch).toEqual(original.firstWatch);
-		});
-	});
-
-	// =========================================================================
-	// StatsParseError
-	// =========================================================================
-
-	describe('StatsParseError', () => {
-		it('includes cause when provided', () => {
-			const cause = new Error('Original error');
-			const error = new StatsParseError('Parse failed', cause);
-
-			expect(error.message).toBe('Parse failed');
-			expect(error.cause).toBe(cause);
-			expect(error.name).toBe('StatsParseError');
-		});
-
-		it('works without cause', () => {
-			const error = new StatsParseError('Parse failed');
-
-			expect(error.message).toBe('Parse failed');
-			expect(error.cause).toBeUndefined();
-		});
+	it.each([
+		['with cause', new Error('Original error')],
+		['without cause', undefined]
+	] as const)('StatsParseError works %s', (_name, cause) => {
+		const error = new StatsParseError('Parse failed', cause);
+		expect(error).toMatchObject({ message: 'Parse failed', name: 'StatsParseError', cause });
 	});
 });
